refactor(listings): extract OfferLinkingService types + add service interface
Brings `OfferLinkingService` into line with the two engineering-standards
rules it was the lone exception to in the repo: types in a sibling
`*.types.ts` file (per § Type Definitions in Separate Files) and an
`implements I*Service` declaration (per § Service Interface Implementation).

The canonical pattern was already established in the same package by
`CategoryResolutionService` — types at `application/types/<name>.types.ts`,
interface at `application/interfaces/<name>.service.interface.ts`, service
implements it, main barrel re-exports the interface + types + runtime
arrays. This PR migrates `OfferLinkingService` to follow the exact same
shape.

Changes:
- New: `libs/core/src/listings/application/types/offer-linking.types.ts`
  exports `OfferLinkMethodValues` (new runtime array per § Union Types),
  `OfferLinkMethod` (now derived from `(typeof OfferLinkMethodValues)[number]`),
  `OfferLinkingLookups`, and `OfferLinkingResult`.
- New: `libs/core/src/listings/application/interfaces/offer-linking.service.interface.ts`
  exports `IOfferLinkingService`. Imports `OfferFeedItem` via relative
  path `../../domain/types/offer-feed.types` (depth `../..`, mirroring
  `ICategoryResolutionService`'s all-relative-within-package precedent).
- Updated: `application/services/offer-linking.service.ts` drops the
  three inline declarations, imports them from the new files, and
  declares `implements IOfferLinkingService`. Method body unchanged.
- Updated: `application/services/offer-mapping-sync.service.ts` reroutes
  the `OfferLinkingLookups` import from the service file to the new
  types file. `OfferLinkingService` (class) import unchanged.
- Updated: `application/services/__tests__/offer-linking.service.spec.ts`
  same reroute. No test logic changes; existing 5 cases pass unmodified.
- Updated: `libs/core/src/listings/index.ts` adds type-only re-exports
  for `IOfferLinkingService`, `OfferLinkMethod`, `OfferLinkingLookups`,
  `OfferLinkingResult`, plus the value export of `OfferLinkMethodValues`.
  Matches the in-context precedent for `ICategoryResolutionService` at
  lines 42-48 of the same file.

No public-API change. The `OfferLinkingService` constructor, `linkOffer`
signature, and the three pre-existing type names are all preserved. The
#359 main-barrel purity invariant is preserved (the service class is
still excluded from the main barrel; `barrel-purity.spec.ts` continues
to pass). The Symbol token binding (`useExisting: OfferLinkingService`
in `listings.module.ts`) needs no change because `useExisting` resolves
structurally.

Quality gate: pnpm lint 0 errors, pnpm type-check clean, pnpm test
2890 tests pass across 9 packages including the unchanged
`offer-linking.service.spec.ts` and `barrel-purity.spec.ts`.

Closes #665
